reminder_service/reminder_service/requirements.txt
Django>=3.2,<4.0
djangorestframework
django_celery_beat
pymongo
pika